requirements.txt
flask~=1.1.2
gunicorn
python-binance
pandas
datetime
Flask-Mail
Flask-Bootstrap
Flask-Login
Flask-Script
Flask-SQLAlchemy
Flask-WTF
SQLAlchemy
WTForms
Werkzeug
DateTime
email-validator